+++ src/hooks/common/useRequireAuth/index.ts
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import useAuthStore, { AuthLevel } from '../../../store/useAuthStore';
-import useUserStore from '../../../store/userStore';
+import useUserStore from '../../../store/useUserStore';
 
 // TODO 해당 페이지에 진입할 수 있는 최소 authLevel을 적으면, user의 authLevel을 확인하고 비교하여 반환
 // 해당 auth레벨보다 낮은 경우 home으로 naviage 시키거나 접근 불가능 state를 제공
